async reimplementation of NetworkPeerList
previous implementation was based on blocking non-async waiting and used quite inefficient methods of waiting for a payload of specific type. however, it allowed to have quite a convenient user code.

when waiting for a specific payload type, messages of all types were processed until the one with the right type was found and all those other messages were then returned to the incoming queue, but seemingly in a potentially wrong order

this commit implements all thing in async way and a small cost of slightly more complex user code at some places (that's when multiple message types can arrive at arbitrary order)